Fixed the problems
Changed the image paths with constants, updated the constants file and changed the name of the classes and files.
