+++ lib/constants/constants.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 const titleFontSize = 38.0;
 const subtitleFontSize = 24.0;
 const secondSubtitleFontSize = 18.0;
@@ -15,6 +17,9 @@ const registerPageImagePath = "assets/registerpageimage.png";
 const blueLogoPath = "assets/bluelogo.png";
 const blueGradientPath = "assets/bluegradient.png";
 const loginPageImagePath = "assets/loginpageimage.png";
+
+//COLORS
+const buttonColor = Color.fromARGB(255, 42, 157, 143);
 
 //PATTERNS
