+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:buildnow/loginpage/loginpage.dart';
-import 'package:buildnow/register_supplier/suppl_registerpage_suppl.dart';
+import 'package:buildnow/register_supplier/supplier_registerpage.dart';
 import 'package:buildnow/registerpage/registerpage_form.dart';
 import 'package:flutter/material.dart';
 
@@ -24,7 +24,7 @@ class _MyAppState extends State<MyApp> {
 
   Widget build(BuildContext context) {
     return MaterialApp(
-        home: SupplRegisterPage()
+        home: LoginPage()
     );
   }
 }
