+++ lib/register_supplier/supplier_registerpage_logo.dart
@@ -1,10 +1,11 @@
+import 'package:buildnow/constants/constants.dart';
 import 'package:flutter/material.dart';
 
-class SupplRegisterPageLogo extends StatelessWidget {
+class SupplierRegisterPageLogo extends StatelessWidget {
   final screenWidth;
   final screenHeight;
 
-  SupplRegisterPageLogo({this.screenWidth, this.screenHeight});
+  SupplierRegisterPageLogo({this.screenWidth, this.screenHeight});
 
   @override
   Widget build(BuildContext context) {
@@ -12,7 +13,7 @@ class SupplRegisterPageLogo extends StatelessWidget {
       right: 0,
       top: screenHeight * 0.016,
       child: Image.asset(
-        "assets/bluelogo.png",
+        blueLogoPath,
         scale: 1 / screenWidth * 22500
       )
     );
